- Add `sessionFrequency` to `Patient` type and refactor related logic. - Refactor patient and user typing for better type safety. - Update `messageRoutes` and cleanup unused imports in admin components. - Upgrade dependencies and resolve registry sources in `package-lock.json`.
