+++ backend/Readme.md
@@ -1,0 +1,19 @@
+# Backend API Documentation
+
+## Requirements
+
+- Node.js (v18.0.0 or higher)
+- npm (v9.0.0 or higher)
+- MySQL (v8.0 or higher)
+
+## Scripts
+- `npm start:` Inicia el servidor.
+- `npm run dev`: Inicia el servidor en modo desarrollo con nodemon.
+- `npm run setup`: Corre el script inicial de setup.
+- `npm run db:create`: Crea la base de datos (MySQL).
+- `npm run db:migrate`: Aplica las migraciones.
+- `npm run db:seed:all`: Carga los seeders.
+- `npm run db:reset`: Drop + create + migrate + seed (reset total).
+- `npm run lint`: Linta el código y muestra errores/warnings.
+- `npm run lint:fix`: Linta y corrige automáticamente los errores de formato que pueda (espacios, comillas, imports, etc).
+
