+++ backend/config/config.js
@@ -4,7 +4,7 @@ require('dotenv').config();
 module.exports = {
 	development: {
 		username: process.env.DB_USER || 'root',
-		password: process.env.DB_PASS || 'root1234',
+		password: process.env.DB_PASS || '1234',
 		database: process.env.DB_NAME || 'ippl_db',
 		host: process.env.DB_HOST || '127.0.0.1',
 		port: process.env.DB_PORT || 3306,
