+++ backend/package.json
@@ -5,6 +5,8 @@
 	"main": "src/index.js",
 	"scripts": {
 		"start": "node src/index.js",
+        "lint": "eslint src/**/*.js",
+        "lint:fix": "eslint src/**/*.js --fix",
 		"db:create": "sequelize-cli --config config/config.js db:create",
 		"db:migrate": "sequelize-cli --config config/config.js db:migrate",
 		"db:seed:all": "sequelize-cli --config config/config.js db:seed:all",
@@ -30,6 +32,9 @@
 		"uuid": "^11.1.0"
 	},
 	"devDependencies": {
+		"@eslint/js": "^9.34.0",
+		"eslint": "^9.34.0",
+		"globals": "^16.3.0",
 		"nodemon": "^2.0.22",
 		"sequelize-cli": "^6.6.3"
 	}
