+++ backend/src/app.js
@@ -7,7 +7,7 @@ const usersRouter = require('./routes/users');
 const uploadRouter = require('./routes/upload');
 const postsRouter = require('./routes/posts');
 const professionalsRouter = require('./routes/professionals');
-const messagesRouter = require('./routes/messages');
+const messagesRouter = require('./routes/messageRoutes');
 const statsRoutes = require('./routes/stats');
 const contentRoutes = require('./routes/content');
 const statusRequestsRoutes = require('./routes/statusRequests');
